+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+Version **0.1.1**
+
+Date Released: **02/16/2026**
+
+- Database migration to PostgreSQL
+- UI Improvements
+
 Version **0.1.0**
 
-Date Released: **02/15/2026**
+Date Released: **01/05/2026**
 
 - Initial release with sleek frontend and backend storage
 - Fully operational bounding and transcription interface